feat: refact login / register in one auth controller
