+++ lib/pages/auth/login_view.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:go_router/go_router.dart';
+import 'package:tawkie/pages/auth/auth.dart';
 import 'package:tawkie/widgets/layouts/login_scaffold.dart';
 import 'package:tawkie/widgets/matrix.dart';
 
-import 'login.dart';
-
 class LoginView extends StatelessWidget {
-  final LoginController controller;
+  final AuthController controller;
 
   const LoginView(this.controller, {super.key});
